internal: Move denormalize logic to one folder
